# improve performance/ergonomics of reading compound datatypes