When reading compound datatypes this replaces `HDF5Compound` with a named tuple with the correct data layout. The data is directly read into the output buffer instead of going through an `IOBuffer`. This results in very large performance gains. 

I wrote a small [benchmark](https://gist.github.com/kleinhenz/ddb8fa7de366e917c2520b963e0bb1f4) to test the performance of reading a one million element dataset of a small compound datatype with three float fields.

Before
```
BenchmarkTools.Trial:
  memory estimate:  329.44 MiB
  allocs estimate:  9000084
  --------------
  minimum time:     1.689 s (19.72% GC)
  median time:      1.878 s (24.72% GC)
  mean time:        1.860 s (25.70% GC)
  maximum time:     2.013 s (31.63% GC)
  --------------
  samples:          3
  evals/sample:     1⏎
```

After:
```
BenchmarkTools.Trial:
  memory estimate:  22.89 MiB
  allocs estimate:  64
  --------------
  minimum time:     4.305 ms (20.43% GC)
  median time:      4.584 ms (20.69% GC)
  mean time:        4.717 ms (23.21% GC)
  maximum time:     53.451 ms (91.03% GC)
  --------------
  samples:          1059
  evals/sample:     1⏎
```

Additionally I fixed support for variable length strings as fields of compound datatypes which seemed to be broken before.

In the case where the compound type contains strings/arrays/vlen types a copy is made to convert from the hdf5 compatible type (e.g. `Cstring`, `FixedArray`, etc.) to native julia types. This has some performance cost but I think is almost always what you want to do. 

Besides being much much faster I think that getting an array of `NamedTuples` is much easier to work with than the `HDF5Compound` struct there was before.

This should close #408. I prefer this solution to #559 since it doesn't change the interface and doesn't require manual specification of the struct.