@musm 2019-11-03
We can certainly make a major version bump. 

> One issue I am currently seeing is that you represent FixedSize Vectors as Vectors, while the "in-memory" representation is a NTuple.

Can you double check here.  The code here is using a a fixed size NTuple not Vectors for FixedArray?