@kleinhenz 2019-11-03
In the `normalize_types` function I translate from `HDF5.FixedArray` which is just a wrapper around an `NTuple` with the shape information saved in the type into an actual `Array`. This is definitely a design decision that I could see going either way. My thought was that in most cases you will want to work with the data as an array so it makes sense to just translate. If we depended on `StaticArrays` we could just use that, but that is probably overkill for this package. 

The reason I decided to do the type translation here is because it is basically necessary for string types. You don't want to return named tuples with `Cstring` members which point to memory allocated by hdf5. So if in order to get sensible julia types we have to do some type translations for strings I thought we might as well also do it for arrays. This could easily be changed if having compatibility with existing structs is more important.

In the case where there are no string or array members then I think using `reinterpret` in order to get to an equivalent struct representation should just work.