@tknopp 2019-11-04
Can we make the type normalization optional? Maybe with the current behavior being the default.