@kleinhenz 2019-11-04
In order to make it optional it seems like you need to either add a new global flag which is a bit unappealing or else add a new mechanism for passing keyword arguments through all of the machinery which seems like sort of a big change just for this. I can't just add it to the relevant method because usually this will be invoked through `h5read` or `read(dset)`.

What would be the advantage of using this and then reinterpreting the data into your custom struct over just reading directly into the custom struct?

IMHO write support is separate and maybe out of scope. I guess `JLD` already solves the problem of how to map julia types to compound hdf5 types. It's probably best to leave general serialization logic there.