@kleinhenz 2019-11-19
@musm what do you think about this approach?