@musm 2019-12-04
Yeah I think the approach here is a good one. I need to play around with it a bit more. 

What about wrapping the `NamedTuple` in an `HDF5Compound` container? This might help users more easily upgrade or transition.