@kleinhenz 2019-12-06
How would you want to wrap it? To maintain the old interface or to just provide an outer type? 

My instinct would be to just return the named tuple directly since

1) the old interface was inconvenient enough to justify a breaking change
2) people already know how named tuples work so adapting code to use the named tuples will have less friction than trying to figure out the semantics of a new `HDF5Compound` container.