@musm 2020-03-02
Any chance we could also get this to work with a variant of https://github.com/JuliaIO/HDF5.jl/pull/341 in the future?