@kleinhenz 2020-03-02
yeah I think that should work fine. I already checked that this integrates with the complex support so float16 should work the same way. One thing that doesn't work currently is hyperslab support since we have an annoying check for `T <: HDF5Scalar` in the `getindex` code. I think a lot of that code should probably be reworked to be more generic.